Modified upload logic;

+++ modules/Feishu/Feishu_messages.py
@@ -136,6 +136,8 @@ class FeishuMessageHandler(FeishuApp):
 
         # 处理业务结果
         stored_path = store_path / response.file_name
+        if stored_path.exists():
+            return False
         f = open(stored_path, "wb")
         f.write(response.file.read())
         f.close()
